requirements.txt
streamlit
pandas
python-dotenv
langchain>=0.1.13
langchain-community>=0.0.20
langchain-google-genai>=0.0.6
google-generativeai>=0.3.2
faiss-cpu